working version silhouette

+++ framework4/main.c
@@ -194,13 +194,10 @@ ray_trace(void)
         for (i = 0; i < framebuffer_width; i++)
         {
             // compute u and v
-            int u = image_plane_width * (((i + 0.5)/(float)framebuffer_width) - 0.5);
-            int v = -image_plane_height * (((j + 0.5)/(float)framebuffer_height) - 0.5);
+            float u = image_plane_width * (((i + 0.5)/(float)framebuffer_width) - 0.5);
+            float v = -image_plane_height * (((j + 0.5)/(float)framebuffer_height) - 0.5);
             
-            // compute ray direction (perspective)
-/*            vec3 ray_direction = {forward_vector.x + u * right_vector.x + v * up_vector.x, */
-/*                                  forward_vector.y + u * right_vector.y + v * up_vector.y,*/
-/*                                  forward_vector.z + u * right_vector.z + v * up_vector.z};*/
+            // compute ray direction
             vec3 pixel_vector = v3_add(v3_multiply(right_vector, u), v3_multiply(up_vector, v));                   
             vec3 ray_direction = v3_add(forward_vector, pixel_vector);
